basic newPage structure.
