+++ client/src/pages/loginPage/components/Login.jsx
@@ -34,10 +34,10 @@ const Login = props => {
 
   return (
     <Flex
-    flexDir="column"
-    mb="2"
-    justifyContent="center"
-    alignItems="center"
+      flexDir="column"
+      mb="2"
+      justifyContent="center"
+      alignItems="center"
     >
       <Stack
         flexDir="column"
